refactor(bingimages): moved dompaths in a separate file
