+++ src/search/engines/bingimages/options.go
@@ -13,46 +13,6 @@ var Info = engines.Info{
 	ResultsPerPage: 35,
 }
 
-type thumbnailDomPaths struct {
-	Path   string
-	Height string
-	Width  string
-}
-
-type bingImagesDomPaths struct {
-	Result       string
-	Metadata     string
-	Title        string
-	ImgFormatStr string
-	Thumbnail    []thumbnailDomPaths
-	Source       string
-}
-
-var dompaths = bingImagesDomPaths{
-	Result:       "ul.dgControl_list > li",
-	Metadata:     "a.iusc", // e.Attr("m")
-	Title:        "div.infnmpt > div > ul > li > a",
-	ImgFormatStr: "div.imgpt > div > span",
-	Thumbnail: []thumbnailDomPaths{
-		{
-			Path:   "a.iusc > div > img.mimg",
-			Height: "height",
-			Width:  "width",
-		},
-		{
-			Path:   "a.iusc > div > div > div.mimg > div",
-			Height: "data-height",
-			Width:  "data-width",
-		},
-		{
-			Path:   "a.iusc > div > div > div.mimg > img",
-			Height: "height",
-			Width:  "width",
-		},
-	},
-	Source: "div.imgpt > div > div.lnkw > a",
-}
-
 var Support = engines.SupportedSettings{
 	Locale: true,
 }
